requirements.txt
networkx==3.4.2
python-dotenv==1.0.1
pyyaml==6.0.2
neo4j==5.28.1
numpy==2.2.6
